requirements.txt
# requirements.txt

Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
google-generativeai==0.5.4

# Packages for the worker
celery==5.4.0
redis==5.0.4